requirements.txt
colorama==0.4.6
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.0
pip==22.3.1
pluggy==1.0.0
pytest==7.3.0
setuptools==65.5.1
tomli==2.0.1
wheel==0.38.4
selenium
webdriver-manager==3.8.6
allure-pytest